requirements.txt
slither-analyzer
click
